network: include Chunk Validators in Tier1 (#13527)
Adds chunk validators to the Tier1 participant set. In mainnet, T1 will
grow in size from 100 to 250 participants.

Changing the function which calculates the participant set should be
sufficient to implement this change. The code in the network crate which
initiates outbound connections and accepts incoming ones operates
according to the calculated participant set.

-----

OptimisticBlock was being broadcast over Tier1 to all peers. The CVs are
not interested in receiving OB, so the behavior is changed to send only
to BPs.**

The following message types sent over T1 are not affected in any way:
- `Tier1Handshake`, `HandshakeFailure`, and `LastEdge`, are used only to
manage the connection.
- `BlockApproval` is routed specifically to the next block producer.
- `VersionedChunkEndorsement` is routed specifically to the next few
block producers.
- `VersionedPartialEncodedChunk` is routed to the block producers of
current and next epochs.
- `ContractCodeRequest` is routed to one randomly chosen chunk producer.
- `ContractCodeResponse` is routed back to the author of the incoming
request.

The following routed message types were already allowed on T1 and will
now be sent over T1 more consistently:
- `PartialEncodedStateWitness` is routed in parts to all chunk
validators.
- `PartialEncodedStateWitnessForward` is routed to all chunk validators.
- `ChunkContractAccesses` is routed to the chunk validators which aren't
producers.


**I tested out sending the OB only to the next 5 block producers instead
of sending it to all of them, but it caused
`tests::optimistic_block::test_optimistic_block` to fail.