[resharding] Move forks handling logic to resharding_actor (#13525)
Apparently the changes in flat_storage_resharder.rs are so much that
github refuses to identify me moving the file.

Please review
https://github.com/near/nearcore/pull/13525/files/d9dc17c5ae3da771bb9834c1ab29cd18144570ca
for better diff management (first 4 commits).

---

This PR started with the intention to move forks handling logic to a
better location so that we could share it between flat storage resharder
and a future upcoming trie state resharder, but while working on the PR,
a lot of other things just happened. It wasn't possible to split this
into smaller PRs.

I figured the best place to actually put this logic isn't the resharding
manager but rather the resharding actor as we can very easily utilize
the single thread-ness of the actor to schedule backoffs till some
resharding block becomes final. That's the most natural place to have
that.

Anyway, continuing with the story... it wasn't easy to extract out the
logic and it was super complicated and very involved to begin with. In
effect, we had a very convoluted logic and interaction between the flat
storage resharder and the resharding manager which I'll try to explain
below. This now get's a lot simplified :)

Change in component hierarchy
- ReshardingManager is a member of chain.
- ReshardingActor existed independently outside chain and handles long
running resharding tasks.
- [new] FlatStorageResharder is now a member of ReshardingActor and not
ReshardingManager.
- [new] FlatStorageResharder now no longer has the sender for
ReshardingActor (as it's simply a component of the actor now) but
instead the resharding manager calls the resharding actor to start
resharding.

Before logic
- ReshardingManager calls `flat_storage_resharder.start_resharding`
- FlatStorageResharder sends `FlatStorageSplitShardRequest` message to
ReshardingActor by with `resharder` i.e. itself as a parameter, which is
kinda insane.
- ReshardingActor simply takes the resharder and calls
`resharder.split_shard_task`. There's a layer of forks checking here by
waiting on `resharder.split_shard_check_readiness`.
- `split_shard_task` in FlatStorageResharder again sends
`FlatStorageShardCatchupRequest` message to ReshardingActor via the
sender.
- ReshardingActor calls `resharder.shard_catchup_task`. This is finally
where we finish the splitting of flat storage.


After logic
- ReshardingManager sends `ScheduleResharding` message to the
ReshardingActor for every resharding block.
- ReshardingActor gathers all `ScheduleResharding` requests and waits
till some resharding block becomes final
- Once a resharding block is final, actor calls
`flat_storage_resharder.start_resharding()` as a long running task. Fin.

I tried to replicate as many of the flat storage manager unit tests as I
could but they are quite extensively covered by resharding_v3 testloop
tests, so I'm not super concerned about them.

Resume resharding util simply creates a new instance of
flat_storage_resharder and calls resume function. In the future with
trie state resharder, we can similarly call
`trie_state_resharder.resume(..)`

`MemtrieReloadRequest` was a noop, just removed it. We can add in back
later as per requirement.