feat: add rpc query for viewing global contract code (#13547)
This PR adds `QueryRequest` for retrieving global contract code. This is
needed because `ViewCode` query only allows retrieving contracted used
as part of the specified accounts. Global contracts deployed separately,
so we need to have a dedicated view query.

2 separate enum variants are introduced for queries by hash
(`ViewGlobalContractCode`) and by account id
(`ViewGlobalContractCodeByAccountId`) to have simple API and keep it
consistent with `DeployGlobalContract`/`UseGlobalContract` actions.

`QueryResponseKind::ViewCode` is reused for the response.

Also `QueryError::NoGlobalContractCode` error is introduced, similar to
`NoContractCode`.

Closes #13531.