[resharding] Increment refcount for child state col (#13566)
This PR is part of a multi change sequence of PRs for the proper fix for
refcount issue in testnet/mainnet.

With this PR we are changing the shard_uid mapping to write data to both
parent state column and child state column.

As part of the PR, we are disabling state column GC related tests. These
would be enabled once we have the other part merged into master as well,
i.e. copying the base state from the parent to the children.