refactor: move state witness creation from Client to ChainStore (#13607)
Prerequisite for debug tool to recreate state witness. It will be
located in state-viewer, and we don't want to create Client and even
Chain for debug purposes. It is more than enough to have ChainStore and
EpochManager.

GitHub doesn't recognise `git mv` is the original file is still present,
so here
https://github.com/near/nearcore/commit/4904c596d7081ed48af91c06e9bd98035345eb39
you can make sure that the logic is simply moved with minimal changes.