refactor: Return CryptoHash instead of &CryptoHash from chunk headers. (#13653)
CryptoHash implements Copy trait so returning value shouldn't be much
more expensive compared to reference. Additionally most usages of
methods are from ShardChunkHeaderInner which are dereferenced in
implementations of ShardChunkHeader anyway.

Returning references makes it hard to remove fields from a new version
of the shard header inner since we cannot return references to temporary
values. (For spice we plan to introduce a new version with fields
removed from chunk header and using defaults until all use cases are
properly migrated. With an alternative of using unreachable! there's a
higher risk of accidentally crashing the node.)