[test] Include nayduck nightly tests in nayduck CI (#13659)
This PR does a couple of things
- Make ci.txt the default nayduck test file (DEFAULT_TEST_FILE)
- Include all the tests in nightly.txt that were NOT in ci.txt

More context

Originally we did not run Nayduck with CI on every PR but instead we
used to have nightly Nayduck runs once a day. With the new nayduck CI
framework, it's no longer necessary to have the nightly runs.

Additionally we had the problem that some of the tests were included in
the nightly runs but not in the CI which lead to a blind spot in test
coverage. This was highlighted by the zulip thread [#general > nayduck @
💬](https://near.zulipchat.com/#narrow/channel/295302-general/topic/nayduck/near/521738339)
where we failed to keep track of
`ultra_slow_test_all_chunks_accepted_1000` test as it was only run in
nightly.

TODO
- Change nayduck to no longer run nightly jobs