refactor: return references to CryptoHash from chunk getters. (#13664)
In some cases for CryptoHash returning a reference may avoid unnecessary
copy of the data so returning a reference would either as efficient or
more efficient as returning a copy from getters. (As far as I understand
it may be as efficient if call will be inlined.)

Since ChunkHash is a wrapper of CryptoHash I included it in this PR as
well. It's very similar to the CryptoHash, but doesn't implement Copy
trait.

Additionally using getters with references makes things more consistent
and can highlight some unnecessary copies on the caller's side.

There was some related discussion here:
https://github.com/near/nearcore/pull/13657#issuecomment-2941546964