prepare: duplicate prepare v2 and create v3 (#13602)
Much like the duplicate of nearvm(2), this is a simple duplicate of the
old prepare v2 code. The new code is only invoked when the
`reftypes_bulk_memory` is enabled. This will replace the old code once
the protocol version upgrade goes through. At which point we will retain
just one implementation of preparation again.

No functional changes at this point in time. The prepare v3 code may
have subtle changes due to upgrades to wasmparser and finite-wasm v0.6.
This is ultimately the reason why this duplication is happening at all.

Built on top of https://github.com/near/nearcore/pull/13581

---------

Co-authored-by: Vlad Frolov <frolvlad@gmail.com>