[shard_tracker] 2/n Simplify cares_about_shard_this_or_next_epoch (#13697)
With PR https://github.com/near/nearcore/pull/13696 we added the
validator_signer to the shard_tracker. In this PR we are simplifying the
calls to `cares_about_shard_this_or_next_epoch` to not pass `account_id`
and `is_me` as parameter.

There's a new function
`cares_about_shard_this_or_next_epoch_for_account_id` that is specific
to answering queries with an account_id provided (and is_me as false).