[shard_tracker] 4/n Simplify cares_about_shard (#13699)
Follow up on https://github.com/near/nearcore/pull/13696, and this time
working with the `cares_about_shard` function and simplifying it.

Note that with this change we lose a bunch of `me: Option<AccountId>`
parameters being passed around, which is a good sign that things are
moving in the right direction.

A couple of actors/modules like ChunkExecutorActor and ViewClientActor
lose the `validator_signer`, which is also a good sign of
simplification.