[chain] 5/n Add validator_signer to chain (#13700)
In PR https://github.com/near/nearcore/pull/13699 we identified an
unused parameter `validator: MutableValidatorSigner` being passed to the
chain. This was not a member of the chain but it makes sense to have
this as a member to get the account_id.

Based on this, we can simplify the parameters passed in a lot of places,
which we would do as part of the next PR.