feat: spice: implement witness creation and validation. (#13905)
As part of this PR I add witness creation and validation together with
not recording execution results unless execution is endorsed.

With spice we have to change witness to include chunk header of the
chunk which execution we are validating (not the next one as is done
without spice). Because of this some of the logic couldn't be reused.

I didn't figure out a nice way to break this into a smaller PRs
unfortunately, but I still left TODOs for logic I thought not essential
for now.

Part of https://github.com/near/nearcore/issues/13387