feat: add adversarial feature to skip chunk production window (#13953)
This PR implements #13912 as part of adversarial behavior so we can have
it in master.

New `AdvProduceChunksMode::SkipWindow` is added to allow skipping a
sequence of `skip_length` chunks in a row deterministically randomly
chosen in a window of size `window_size`.

This allows us to test that network can recover after a long sequence of
missing chunks.