state sync: measure time spent to validate parts (#13944)
Addresses a gap in our metrics. We measure time to generate parts, time
to transfer them, and time to apply them. This validation step is
separate from the full application process and also takes a non-trivial
amount of time.

Did state sync on a testnet node with this code and observed that
received parts typically take .5-1s to validate.