[memtrie] Introduce snapshot in memtrie (#13933)
This PR introduces the concept of a snapshot for memtries. This is
useful if we want to save the state at a specific height, for example in
the case of state_sync.

Conceptually it's very simple, where we are simply increasing the
refcount of the state_root of the snapshot height so that it doesn't get
garbage collected.

Caveats: Saving a snapshot for an extended period of time can lead to
extra memory consumption as the nodes corresponding to the snapshot
height are not cleaned up. This is something that the user of the
snapshot should be careful with.