apply-range: automatically process all shards in parallel (#13971)
Not only does this make the "regular" use of apply-range faster, it also
works as a more realistic benchmark, since we never really apply chunks
in isolation.

The implementation is pretty naive – it just goes as fast as possible
for each individual shard. No synchronization or barriers between shards
here.

The original behaviour is still available by passing in `--shard-id` if
needed.