[actix_migration] Change Handler trait to be independent of actix::Message (#13985)
This change is required for the future ground work of implementing a
tokio based actor implementation as highlighted in poc
https://github.com/near/nearcore/pull/13981

`Handler` trait now takes both Message and Response as template type
with the default response type being `()`.

If we implement `Handler<M, R>` with non-default response type, we can
call `send_async` from the sender and wait for the future, while if we
implement `Handler<M>` with default response type, we can call the sync
`send` method from the sender.

The individual message types still continue to derive `actix::Message`
as that's required till we migrate over all actors to tokio.