[actix_migration] Introduce tokio based actor (#13987)
This PR introduces Tokio based actor which is the alternative and
eventual replacement of the Actix based actor.

We introduce the `TokioRuntimeHandle<Actor>`, which is the equivalent of
`actix::Addr<Actor>`.

The `TokioRuntimeHandle` is a low cost cloneable struct that can be used
to get the sender and the async spawner for the actor.