feat: spice: Persist spice core data. (#14001)
Closes https://github.com/near/nearcore/issues/13388

In this PR I add persistence for data we need to store for spice core
and related additional validations to make sure we don't store garbage
that cannot be garbage collected.

I also introduced SpiceEndorsementWithSignature and started to use it in
Endorsement for spice block to make code more consistent with what is
stored.